test: isolate runtime contract doc parity
