+++ lib/index.ts
@@ -2,7 +2,6 @@ export * from "./accounts.js";
 export * from "./storage.js";
 export * from "./config.js";
 export * from "./constants.js";
-export * from "./runtime-contracts.js";
 export * from "./types.js";
 export * from "./logger.js";
 export * from "./auth/auth.js";
